# openapi.action.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GitHub Agent Gateway",
    "version": "1.0.0",
    "description": "Safely manage allowlisted GitHub repositories. Preview code changes before applying them. Destructive and admin operations require exact confirmation strings and server-side feature flags."
  },
  "servers": [
    {
      "url": "https://YOUR-WORKER.workers.dev"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/repos/{owner}/{repository}/releases": {
      "get": {
        "summary": "List releases",
        "operationId": "listReleases",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Create a draft or published release",
        "operationId": "createRelease",
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tagName"
                ],
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/v1/repos/{owner}/{repository}/releases/latest": {
      "get": {
        "summary": "Get latest release",
        "operationId": "getLatestRelease",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/repos/{owner}/{repository}/releases/{releaseId}": {
      "get": {
        "summary": "Get a release",
        "operationId": "getRelease",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "releaseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "summary": "Update a release",
        "operationId": "updateRelease",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "releaseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a release",
        "operationId": "deleteRelease",
        "responses": {
          "204": {
            "description": "Success"
          }
        },
        "description": "Disabled by default. Confirmation: DELETE RELEASE owner/repository releaseId.",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "releaseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Confirmation"
              }
            }
          }
        }
      }
    },
    "/v1/repos/{owner}/{repository}/tags": {
      "get": {
        "summary": "List tags",
        "operationId": "listTags",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Create a lightweight or annotated tag",
        "operationId": "createTag",
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tag",
                  "targetSha"
                ],
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/v1/repos/{owner}/{repository}/tags/{tag}": {
      "delete": {
        "summary": "Delete a tag",
        "operationId": "deleteTag",
        "description": "Disabled by default. Confirmation: DELETE TAG owner/repository tag.",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Confirmation"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Confirmation": {
        "type": "object",
        "required": [
          "confirmation"
        ],
        "properties": {
          "confirmation": {
            "type": "string"
          }
        }
      },
...
    }
  }
}